requirements.txt
celery==4.4.6
Django==3.0.8
django-filter==2.3.0
djangorestframework==3.11.0
Markdown==3.2.2
pip==20.1.1
setuptools==49.2.0
wheel==0.34.2
